+++ tests/content_handler_context/ctx_box.js
@@ -1,0 +1,5 @@
+class Ctx_Box extends Jqhtml_Component {
+  on_own_click() {
+    this.own_click_this = this;
+  }
+}